backend/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database (Optional - Commented out for local install on Python 3.14)
# sqlalchemy>=2.0.23
# psycopg2-binary>=2.9.9
# alembic>=1.13.0

# Machine Learning
numpy>=1.24.4
scikit-learn>=1.3.2
pandas>=2.1.3

# Image Processing
Pillow>=10.1.0
opencv-python>=4.8.1.78
scikit-image>=0.22.0

# Explainable AI
lime>=0.2.0.1
matplotlib>=3.8.2

# Configuration and Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# HTTP and API
requests>=2.31.0
httpx>=0.25.2

# System Monitoring
psutil>=5.9.6

# Development and Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# Logging
structlog>=23.2.0

# Utilities
tqdm>=4.66.1
python-json-logger>=2.0.7
huggingface_hub